requirements.txt
SpeechRecognition
pyserial
pyaudio